requirements.txt
python3-pip
dnspython